Refactor: move the logic out of the main.py file
